Create `OffenderDetailService`
This commit creates an `OffenderDetailService`, extracting functions from `OffenderService` that returns `PersonInfoResults`.

The purpose of this split of the `OffenderService` is to reduce the amount of duplicated unit testing we have around LAO logic. The functions in `OffenderDetailService` delegate to already well-tested functions in `OffenderService` to retrieve offender info (including LAO), so it’s redundant for those function calls to also be testing that logic. Whilst the tests for `OffenderDetailService` are still testing this, in a subsequent commit they will not and instead mock response from OffenderService` will be used.

Logically it may make more sense for this to be called `OffenderService` and the existing `OffenderService` be called `OffenderSummaryService`. We can make that change in a future commit.
