+++ src/test/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/unit/service/cas1/Cas1ApplicationCreationServiceTest.kt
@@ -58,8 +58,8 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.prisonsapi.InmateS
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.results.AuthorisableActionResult
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.results.CasResult
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.AssessmentService
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.OffenderDetailService
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.OffenderRisksService
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.OffenderService
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.Cas1ApplicationCreationService
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.Cas1ApplicationCreationService.Cas1ApplicationUpdateFields
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.Cas1ApplicationDomainEventService
@@ -76,7 +76,6 @@ import java.util.UUID
 @SuppressWarnings("LargeClass")
 class Cas1ApplicationCreationServiceTest {
   private val mockApplicationRepository = mockk<ApplicationRepository>()
-  private val mockOffenderService = mockk<OffenderService>()
   private val mockOffenderRisksService = mockk<OffenderRisksService>()
   private val mockAssessmentService = mockk<AssessmentService>()
   private val mockOfflineApplicationRepository = mockk<OfflineApplicationRepository>()
@@ -92,10 +91,10 @@ class Cas1ApplicationCreationServiceTest {
   private val mockLockableApplicationRepository = mockk<LockableApplicationRepository>()
   private val mockCas1CruManagementAreaRepository = mockk<Cas1CruManagementAreaRepository>()
   private val mockCas1OffenderService = mockk<Cas1OffenderService>()
+  private val mockOffenderDetailService = mockk<OffenderDetailService>()
 
   private val applicationService = Cas1ApplicationCreationService(
     mockApplicationRepository,
-    mockOffenderService,
     mockOffenderRisksService,
     mockAssessmentService,
     mockOfflineApplicationRepository,
@@ -112,6 +111,7 @@ class Cas1ApplicationCreationServiceTest {
     mockLockableApplicationRepository,
     mockCas1CruManagementAreaRepository,
     mockCas1OffenderService,
+    mockOffenderDetailService,
   )
 
   @Nested
@@ -997,7 +997,7 @@ class Cas1ApplicationCreationServiceTest {
       every { mockObjectMapper.writeValueAsString(defaultSubmitApprovedPremisesApplication.translatedDocument) } returns "{}"
       every { mockApplicationRepository.findByIdOrNull(applicationId) } returns application
       every { mockApplicationRepository.save(any()) } answers { it.invocation.args[0] as ApplicationEntity }
-      every { mockOffenderService.getInmateDetailByNomsNumber(any(), any()) } returns AuthorisableActionResult.Success(
+      every { mockOffenderDetailService.getInmateDetailByNomsNumber(any(), any()) } returns AuthorisableActionResult.Success(
         InmateDetailFactory().withCustodyStatus(InmateStatus.OUT).produce(),
       )
